requirements.txt
django>=3.2,<4.0
gunicorn
requests
psycopg2-binary
django-environ
django-cors-headers
djangorestframework
pillow==9.3.0
djangorestframework-simplejwt==5.2.2
beautifulsoup4
lxml

